--- Assets/Scripts/CutGrass.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2865b261c84370d4abf235316e285c46
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
