+++ Assets/StreamingAssets/BGM.bank.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: a292bbd06231aa847aa151ce605eaf3d
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
