+++ Assets/StreamingAssets/Master.bank.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5f9f21a4970404e419eb6c60321570b6
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
