+++ Assets/StreamingAssets/Master.strings.bank.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ec624e56d8b848046935038738997a86
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
