+++ Assets/StreamingAssets/SFX.bank.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 605f2fa388de97842872ed6407f53a56
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
